Revert "Moved classes for sake of readability"
This reverts commit c207c084e6718fa6f4567ca465ca0241cd7127df.
