+++ src/index.ts
@@ -52,7 +52,7 @@ import {
   IIndex,
   Index,
 } from './index-class';
-import {Query} from './exported/query';
+import {Query} from './query';
 import {
   DatastoreRequest,
   CommitCallback,
