+++ test/request.ts
@@ -25,7 +25,7 @@ import {PassThrough, Transform} from 'stream';
 import {google} from '../protos/protos';
 import * as ds from '../src';
 import {entity, Entity, KeyProto} from '../src/entity.js';
-import {IntegerTypeCastOptions, Query, QueryProto} from '../src/exported/query';
+import {IntegerTypeCastOptions, Query, QueryProto} from '../src/query.js';
 import {outOfBoundsError} from './entity';
 import {
   AllocateIdsResponse,
